refactor: linearNoConfusionType: use ULift, not `PUnit →`
This PR refactors the juggling of universes in the linear
`noConfusionType` construction: Instead of using `PUnit.{…} → ` in the
to get the branches of `withCtorType` to the same universe level, we use `ULift`.

It seems to work in the simple test case in `noConfusionLinear`.

It does not solve the issue in #8962 (I hoped that lean would handle the
`max` better than the `imax`), but it fails with
```
error: invalid universe level, max (max (max (u+1) (u_1+1)) (v+1)) (imax u v) is not greater than 0
```
in
```
decLevel
```