feat: use synthetic metavariables for autoparams
This PR modifies how autoparams elaborate. Previously they would be synthetic opaque metavariables, which would prevent unification, but now they are synthetic, which merely avoids unification. The `runTactic` function now checks to see whether or not the initial goal is assigned, and if it is it now runs the tactic and then does a defeq check, just like in typeclass inference.

We may consider having non-synthetic-opaque tactic metavariables skip running the tactic entirely if the metavariable is assigned, but it's not clear this would be correct — possibly unification will only partially assign the autoparam.