fix: adjust unsafe trick for upcoming optimization
This PR fixes an unsafe trick where a sentinel for a hash table of
Exprs (keyed by pointer) is created by constructing a value whose
runtime representation can never be a valid Expr. The value chosen
for this purpose was Unit.unit, which violates the inference that
Expr has no scalar constructors. Instead, we change this to a
freshly allocated Unit × Unit value.