fix: use withIncRecDepth in SizeOf deriving
This PR addresses a lean crash (stack overflow) with nested induction
and the generation of the `SizeOf` spec lemmas.

It does not address the underlying issue that in these cases, the
generated SizeOf code does not match the the SizeOf function found by
instance search, and thus the generation fails.