perf: update type information in cases arms for RC optimizations
The `isRef` check being removed here used to be an optimization,
because this structure only tracked whether ref counting operations
need to be inserted at all. Now the structure also tracks whether
the value needs to be checked for being a scalar or not, which is
something that can be refined by a `cases` arm, since inductive
types can have a mix of scalar and non-scalar constructors.