feat: custom `structure` constructors can update binder kinds of parameters
This PR adds a feature where `structure` constructors can override the inferred binder kinds of the type's parameters. For example,
```lean
structure WithLp (p : Nat) (V : Type) where toLp (p) ::
  ofLp : V
```
causes `p` to be an explicit parameter to `WithLp.toLp`. This reflects the feature added in #7742 to be able to override binder kinds for projections.

Closes #9072.