fix: show kernel diagnostics even for examples
This PR surfaces kernel diagnostics even in `example`.

The problem was that the kernel checking happens asynchronously. We cannot use `reportDiag` in `addDecl`, which spawns that task, due to the module hierarchy. For non `example`-declaration, `reportDiag` is called somewhere else later, but for `example`, the `withoutModifyingEnv` in `elabMutualDef` hid the kernel diagnostics. (But only the kernel diagnostics; they are in the `Environment`, while the others are in the `State`).

I also observed that the `reportDiag` in `elabAsync` (but not in `elabSync`) duplicated the reporting, so without `elab.Async true` you get the message twice. To fix this, `reportDiag` now resets the diagnostics. This should avoid reporting counts twice in general (at least within a linear use of the state).