chore: remove tests that recompile builtin definitions
In the early days of the new compiler, it was common to make tests
that manually compiled a definition with the new compiler. The arity
reduction pass in LCNF deliberately does not compute a fixed point
to find a minimal set of used parameters for performance reasons,
but running it a second time can lead to different decisions being
made and a decl arity mismatch. This has been an issue for multiple
people during development. Removing the tests fixes the problem.

Fixes #9186.