fix: make `zero`/`succ` hoverable in `induction`/`cases`
This PR restores the feature where in `induction`/`cases` for `Nat`, the `zero` and `succ` labels are hoverable. This was added in #1660, but broken in #3629 and #3655 when custom eliminators were added. In general, if a custom eliminator `T.elim` for an inductive type `T` has an alternative `foo`, and `T.foo` is a constant, then the `foo` label will have `T.foo` hover information.